audit: Expose LogEntry & LogEntries types publicly.
This requires moving these types to out of the 'audit::commands' module
as it is limited to `pub (crate)` scope.
